third_party/WebKit/LayoutTests/fast/css-grid-layout/named-grid-line-get-set.html

<!DOCTYPE HTML>
<html>
<head>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<link href="resources/grid.css" rel="stylesheet">
<style>
.grid {
    /* Give an explicit size to the grid so that percentage grid tracks have a consistent resolution */
    width: 800px;
    height: 600px;
}
.gridItem {
    grid-column: 1;
    grid-row: 1;
    height: 22px;
    width: 77px;
}
.gridWithFixed {
    grid-definition-columns: "first" 10px;
    grid-definition-rows: "first" 15px;
}
.gridWithPercent {
    grid-definition-columns: 53% "last";
    grid-definition-rows: 27% "last";
}
.gridWithAuto {
    grid-definition-columns: "first" auto;
    grid-definition-rows: auto "last";
}
.gridWithMinMax {
    grid-definition-columns: "first" minmax(10%, 15px);
    grid-definition-rows: minmax(20px, 50%) "last";
}
.gridWithFixedMultiple {
    grid-definition-columns: "first" "nav" 10px "last";
    grid-definition-rows: "first" "nav" 15px "last";
}
.gridWithPercentageSameStringMultipleTimes {
    grid-definition-columns: "first" "nav" 10% "nav" 15% "last";
    grid-definition-rows: "first" "nav2" 25% "nav2" 75% "last";
}
</style>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div class="grid gridWithFixed" id="gridWithFixedElement"></div>
<div class="grid gridWithPercent" id="gridWithPercentElement"></div>
<div class="grid gridWithAuto" id="gridWithAutoElement"></div>
<div class="grid gridWithAuto" id="gridWithAutoWithChildrenElement">
    <div class="gridItem"></div>
</div>
<div class="grid gridWithMinMax" id="gridWithMinMax"></div>
<div class="grid gridWithFixedMultiple" id="gridWithFixedMultiple"></div>
<div class="grid gridWithPercentageSameStringMultipleTimes" id="gridWithPercentageSameStringMultipleTimes"></div>

<script src="resources/grid-definitions-parsing-utils.js"></script>
<script>
    description('Test that setting and getting grid-definition-columns and grid-definition-rows works as expected');

    debug("Test getting grid-definition-columns and grid-definition-rows set through CSS");
    testGridDefinitionsValues(document.getElementById("gridWithFixedElement"), "first 10px", "first 15px");
    testGridDefinitionsValues(document.getElementById("gridWithPercentElement"), "424px last", "162px last");
    testGridDefinitionsValues(document.getElementById("gridWithAutoElement"), "first 0px", "0px last");
    testGridDefinitionsValues(document.getElementById("gridWithAutoWithChildrenElement"), "first 77px", "22px last");
    testGridDefinitionsValues(document.getElementById("gridWithMinMax"), "first 80px", "300px last");
    testGridDefinitionsValues(document.getElementById("gridWithFixedMultiple"), "first nav 10px last", "first nav 15px last");
    testGridDefinitionsValues(document.getElementById("gridWithPercentageSameStringMultipleTimes"), "first nav 80px nav 120px last", "first nav2 150px nav2 450px last");

    debug("");
    debug("Test getting and setting grid-definition-columns and grid-definition-rows through JS");
    testGridDefinitionsSetJSValues("'first' 18px", "66px 'last'", "first 18px", "66px last", "first 18px", "66px last");
    testGridDefinitionsSetJSValues("'first' 55%", "40% 'last'", "first 440px", "240px last", "first 55%", "40% last");
    testGridDefinitionsSetJSValues("'first' auto", "auto 'last'", "first 0px", "0px last", "first auto", "auto last");
    testGridDefinitionsSetJSValues("'first' min-content", "min-content 'last'", "first 0px", "0px last", "first min-content", "min-content last");
    testGridDefinitionsSetJSValues("'first' max-content", "max-content 'last'", "first 0px", "0px last", "first max-content", "max-content last");
    testGridDefinitionsSetJSValues("'first' minmax(55%, 45px)", "minmax(30px, 40%) 'last'", "first 440px", "240px last", "first minmax(55%, 45px)", "minmax(30px, 40%) last");
    testGridDefinitionsSetJSValues("'first' minmax(22em, max-content)", "minmax(max-content, 5em) 'last'", "first 220px", "50px last", "first minmax(22em, max-content)", "minmax(max-content, 5em) last");
    testGridDefinitionsSetJSValues("'first' minmax(22em, min-content)", "minmax(min-content, 5em) 'last'", "first 220px", "50px last", "first minmax(22em, min-content)", "minmax(min-content, 5em) last");
    testGridDefinitionsSetJSValues("'first' minmax(min-content, max-content)", "minmax(max-content, min-content) 'last'", "first 0px", "0px last", "first minmax(min-content, max-content)", "minmax(max-content, min-content) last");
    testGridDefinitionsSetJSValues("'first' 'nav' minmax(min-content, max-content) 'last'", "'first' 'nav' minmax(max-content, min-content) 'last'", "first nav 0px last", "first nav 0px last", "first nav minmax(min-content, max-content) last", "first nav minmax(max-content, min-content) last");
    testGridDefinitionsSetJSValues("'first' 'nav' minmax(min-content, max-content) 'nav' auto 'last'", "'first' 'nav2' minmax(max-content, min-content) 'nav2' minmax(10px, 15px) 'last'", "first nav 0px nav 0px last", "first nav2 0px nav2 15px last", "first nav minmax(min-content, max-content) nav auto last", "first nav2 minmax(max-content, min-content) nav2 minmax(10px, 15px) last");
    testGridDefinitionsSetJSValues("'foo' 'bar' auto 'foo' auto 'bar'", "'foo' 'bar' auto 'foo' auto 'bar'", "foo bar 0px foo 0px bar", "foo bar 0px foo 0px bar", "foo bar auto foo auto bar", "foo bar auto foo auto bar");

    debug("");
    debug("Test getting and setting invalid grid-definition-columns and grid-definition-rows through JS");
    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.display = "grid";
    element.style.gridDefinitionColumns = "'foo'";
    element.style.gridDefinitionRows = "'bar";
    shouldBeEqualToString("getComputedStyle(element, '').getPropertyValue('grid-definition-columns')", "none");
    shouldBeEqualToString("getComputedStyle(element, '').getPropertyValue('grid-definition-rows')", "none");

    element = document.createElement("div");
    document.body.appendChild(element);
    element.style.display = "grid";
    element.style.gridDefinitionColumns = "'foo' 'bar'";
    element.style.gridDefinitionRows = "'bar' 'foo'";
    shouldBeEqualToString("getComputedStyle(element, '').getPropertyValue('grid-definition-columns')", "none");
    shouldBeEqualToString("getComputedStyle(element, '').getPropertyValue('grid-definition-rows')", "none");
</script>
</body>
</html>
